Implement LISA v2 source file converter
Version 2 of the LISA assembler uses a fairly straightforward
tokenized format.  The original CiderPress didn't handle this very
well, so this is all new.  The output generated by LISA's "WRITE"
command, which creates a plain text listing, matches the output of
the converter on the files tested.

All versions of LISA will eventually be handled by this single
converter.

This adds test files for v2, v3, and v4/v5 to the sample set.  One
of the v2 samples was modified to exercise some edge cases.

Note for future explorers: the RANDYS HIRES ROUTINES source file has
some odd control characters embedded in the file, e.g. '\0'.  These
are in fact present in the original sources, and are included in the
text file generated by the WRITE command.  They don't cause the DOS
text file to end early because they're output in high ASCII.