Don't add dirs to xfer list when exporting
The ClipFileSet constructor generates two lists, one for "foreign"
transfers and one for direct transfers.  If the application is in
export mode, we don't want to generate anything in the direct
transfer list, so the remote copy of CP2 won't see anything to do.

We output a list of directories as we work, and weren't checking to
see if we were in export mode.  This caused us to generate a list of
directories on the clipboard.  The pasting side would appear to
succeed, because it had a nonzero set of items, but if the directories
already existed then nothing would appear to happen.

We now suppress generation of directory entries in the direct-xfer
list while in export mode.