First cut at 65xx disassembly converter
The converter generates a disassembly for 8-bit code.  Output
resembles the monitor listing on an Apple //e.  ProDOS 8 calls are
recognized and formatted, but MLI call names are not yet displayed.

The line formatting code is a straight port from the original
CiderPress.

The CPU instruction code and tables were copied from the Asm65 library
in 6502bench.  The code does a lot more than we actually need, but
it required almost no modifications (just a couple of tweaks to pacify
the nullability checker).