Add "rsrc" converter for CLI
The GUI lets you view the formatted contents of the resource fork of
any file, but the CLI didn't have a way to generate it.  This adds
a "rsrc" converter that only formats the resource fork.  This isn't
something that you'd want to see in the GUI, so it's hidden there.