Implement OMF formatter
The converter formats the contents of the OMF segment headers.  This
is different from the original CiderPress, which attempted to
display the segment bodies as well.  That rarely worked well, so now
we just display the various header fields.

The code for doing this was pulled out of 6502bench SourceGen, which
includes an OMF conversion tool.