Implement CP/M text converter
CP/M text files stop at the first Ctrl+Z, which matters when the file
length is specified in 128-byte records.

The trick is figuring out which files are text and which aren't.
The filename extension can help in some situations, but the best
approach seems to be to scan the file contents.  The implementation
here is a direct port from the original CiderPress.

Also, various minor tweaks for consistency and correctness.