Show error when pasting exported files
Files may be transferred between instances of CiderPress II using
drag & drop or copy & paste.  These features can be used for both
add/extract and import/export, but it doesn't really make sense to
export from one copy of CiderPress II and import it into another.
Attempting to do so is likely a mistake.

We weren't allowing it, but silent failures aren't helpful.  We now
show an explicit "don't do that" message on the pasting side.