Replace WPF hack with other WPF hack
This replaces the "dummy element" hack with the "binding proxy"
hack.  This is required to be able to hide DataGrid columns.

The goal was to improve the problems seen in issue #5, but this
doesn't seem to change that behavior.  (Technically, the behavior
has changed: the infinite GC issue isn't happening, and there's a
new series of binding complaints.  But that was true before I
started making changes.)

Yay WPF.