Use custom sort for main file list
The DataGrid sorts individual columns just fine, but sometimes we
want a secondary sort.  For example, if we're asked to sort by file
type, the user is probably looking for specific types of files, so
we should group similar auxtypes together.

This gets a little weird when the set of columns changes, because
we don't clear the sort when e.g. switching from single-directory
to full-file view.  If we switch from ZIP to ProDOS, we could still
be sorting on the Data Format column, which isn't really used for
ProDOS.  We always use pathname as a final sort, so it doesn't look
too bad.

The "reset sort" button is now enabled only when a custom sort is
being used.