Add notes about 65xx disassembly
Also, added REL to the list of 8-bit disasm targets.  Perhaps not
an ideal choice, if the file is full of relocation dictionaries,
but it's probably better than the other options.

Also, Google Groups is dropping Usenet support, so don't mention
groups.google.com as a place to report bugs.