Allow creation of ".dsk" images, sometimes
The CLI create-disk-image command uses the filename extensions to
decide what the disk image format should be.  Creating a new disk
with a name like "foo.dsk" was being blocked because, in the context
of 16-sector 5.25" disk images, it's ambiguous.  In most cases it
means DOS-ordered sectors, but it's often used for ProDOS-ordered
sectors, and it's just generally annoying for emulators to have to
try to guess at what's inside.

It's unambiguous when creating an 800KB 3.5" disk image, or a hard
drive image, so we want to allow it there because many people expect
that to work.

The solution is to reject ".dsk" for disks 320KB and smaller, but
allow it for larger disks.

We now block creation of disks with the ".img" extension, because
it's used in a generic way and isn't closely associated with any one
format.  Before, it was used to create physical-ordered 16-sector
5.25" disk images like those created by a specific version of
Copy ][+, but in practice nobody uses those, so we certainly don't
need to be creating more of them.

(issue #6)