Remove osx-arm64 RID from build
I tried osx-x64 and osx-arm64 self-contained builds on an M1
PowerBook.  The x64 version worked, the arm64 version just printed
"zsh: killed".

This may be due to the "Code Signing" feature added in 11.0.1:
https://developer.apple.com/documentation/macos-release-notes/macos-big-sur-11_0_1-universal-apps-release-notes
"New in macOS 11 on Macs with Apple silicon, and starting in macOS
Big Sur 11 beta 6, the operating system enforces that any executable
must be signed before it’s allowed to run. There isn’t a specific
identity requirement for this signature: a simple ad-hoc signature is
sufficient. [...] This new policy doesn’t apply to translated x86
binaries running under Rosetta 2, nor does it apply to macOS 11
running on Intel-based platforms."

One bug report (https://github.com/dotnet/runtime/issues/49091)
suggests that this can be fixed with "codesign -s - arm64/stuff".
That's a bit awkward because all of the packages are currently
being built on Windows, and that step would require a Mac.

For now, I'm disabling the osx-arm64 package.