Tweak file viewer layout
The file viewer prev/next buttons were centered on the converter
output control, which was fine until a converter with a very large
number of options was used.  The overall window size doesn't change,
so the converter output gets scrunched up, which moves the position
of the prev/next buttons slightly.  If you were trying to move
through a large collection of disparate items, the button might move
out from under your mouse and break your flow.

They're now positioned relative to the window as a whole, slightly
up from the middle.