Improve single-file archive add errors
Attempting to add a file to a single-file archive is prevented at
various levels.  For drag & drop / copy & paste, it was getting
through the worker and into the DiskArc library before it was stopped.
Same for read-only targets.  We now do the appropriate checks in the
GUI app before allowing the operation to begin.