Improve selection order preservation
Change c74e840e corrected the lack of recursive descent for drag and
copy operations, but in doing so caused the list of files to be
sorted.  We want to maintain the order, especially for filesystems
like ProDOS where the order can determine which SYS file is executed
after boot.

We need to sort the list to ensure that directories appear before
their contents.  We could generate directories on the fly when adding
files, but this ordering is critical when doing a recursive delete.

The code now does a more limited sort, shifting the position of
directories so that they appear before the first entry contained
within them.  The selection is not otherwise reordered.