Launch viewer for PDF, RTF, and Word files
Files are identified by extension, HFS file type, and signature.
The file is extracted into a temp file and handed to the default
viewer for that extension.

Windows doesn't allow you to delete an open file, so if the file
is open in the external program when the file viewer is closed, the
temporary file won't be removed.  We can scan for these at launch
and scrub them, ideally with user confirmation since we can't
absolutely guarantee that we created them.

(Issue #7.)