Allow open of read-only files
The retry-as-r/o code was expecting an IOException rather than an
UnauthorizedAccessException.  The CLI was mostly working, because
commands are specifically r/o or r/w and the initial open uses the
appropriate mode, but the GUI was failing.

This adds a read-only test to the CLI test suite.  It wouldn't
have failed before, because it's just detecting success/failure,
but it seems like a useful thing to exercise.

(Issue #9)