Improve handling of PDF/RTF/Word
The file viewer can't clean up its temporary files if an external
viewer still has the file open, so we need to scan for stale temp
files when the program starts.  We're removing "cp2tmp_*" from the
system temporary directory, so it should be safe to remove the files
without user review and confirmation.

The CLI was refusing to export "host image" files, but this seems
wrong.  We now simply extract them as-is.  The GUI was already doing
this.

Fixed a bug where you couldn't export a "host image" file from the
GUI with drag & drop if it was in a file archive.

(Issue #7)