Further improve handling of marginal Pascal disks
If a file is off the end of the disk, display the portion of it
that fits.  This requires trimming the file and marking it as
"dubious" rather than "damaged".  This also applies to files with
bad filenames.

Also, if a file has both IsDubious and IsDamaged set, display '!'
in the CLI catalog listing rather than '?'.