Open read-only if read-write fails
This was always the intention, but it didn't actually work, because
the file sharing argument was still "Read" rather than "ReadWrite".
It needs to be ReadWrite to tell the system that it's okay if the other
process has the file open for writing.

The reason why the file is being opened read-only isn't obvious, so we
want to add a note to the top-level node indicating the problem.

(issue #16)