Bitmap font viewing, part 1
This is for conversion and display of QuickDraw II font files, as
found on the Apple IIgs.  The Macintosh font format is very similar,
but those are stored in individual resources in the system file.
The notes document both, but the converter only handles the IIgs files
because we don't have fancy formatting of individual resources.

The converter is just a stub.  We want to display the glyphs in a
grid, so the Font8x8 class has been pulled in from 6502bench.

Also, updated some comments about font family numbers.