Tweak ProDOS vs. HFS file type selection algorithm
Sometimes ProDOS and HFS types are both available, e.g. ShrinkIt
archives and forked ProDOS files.  We were preferring the HFS type
when displaying disk catalogs, but there are times when we don't
want to do that.  This is somewhat complicated by the fact that,
on an HFS disk, the ProDOS file type is stored by forming the HFS
type in a certain way.

For example:
 - Pointless truetype fonts are FON/$0001, but sometimes they were
   showing as FFIL/DMOV or tfil/movr due to vestigial HFS type info
   in the resource fork.
 - Wolf3D sound files had S16/$0000 embedded in the HFS type in the
   resource fork, but the ProDOS type is different: SND/$0003.

There are cases where we *do* want to use the HFS type, because the
ProDOS type is generic: NON ($00), TXT ($04), and BIN ($06).  If
these files are forked, there's a fair chance they came from a Mac.

The algorithm has been updated so that, if the files live on a
ProDOS disk, the ProDOS type is preferred when it's non-generic.
This seems to produce better results without losing any of the
useful HFS type info.

Note this only applies to how the file attributes are displayed in
the catalog listings, in the CLI and GUI apps.  A "wide" listing
in the CLI always displays both types, as does the GUI "edit
attributes" dialog, and the file converters have access to everything.