Bitmap font viewing, part 3
The font grid seems to be working.  The grid cells are sized to fit
the largest possible glyph, and filled with light gray.  The area
from the origin to the character width is filled with white, and the
glyph is drawn in black.  Any pixels drawn outside the white rectangle
would overlap with previous or subsequent characters.

Cells for characters not represented in the font are filled with
dull red.

I ran a few hundred fonts through it and it seems to be working
correctly.