Add Merlin Assembler import converter
The file format Merlin uses for assembly sources compresses out the
spaces between fields.  To make the lines easier to scan, it flips
the high bits on spaces that are in the middle of comments or quoted
text.  The importer generates the ProDOS form from a plain text file.

The Merlin format allows some flexibility, e.g. a line with only a
label and a comment could be entered:
  LABEL<space>;comment
or:
  LABEL<space><space><space>;comment

The Merlin 16 full-screen editor will store it however you type it.
The assembler uses the presence of the ';' to detect the comment
field, rather than counting separators.

This makes it a little harder to do a simple export+import+compare,
because the excess spaces get crunched out during conversion.
Export+import+export+compare has a similar problem with trailing
spaces after opcodes, which the exporter doesn't try to filter out.