Convert IIgs finder icon files, part 2 (of 2)
Icon files are parsed and displayed in a grid, with one row per icon
entry.  Each row has four columns, for the large/small images and
the associated mask.  The display is patterned after DIcEd, e.g. the
masks are shown with white for the transparent parts and black for the
opaque parts.

The various filters on filename and file types are dumped into the
Notes section.

The ImageGrid class has been updated to allow one or both sets of
labels to be omitted.  Also, the number of columns is no longer fixed
at 16, and the row label can be based on the row number rather than
the item number.  The constructor was getting crowded, so I replaced
it with a Builder class that lets you set the arguments in small
clumps.

The Bitmap8 class has a new FillRect call, for the sake of convenience
and efficiency.