Display Applesoft shape tables
The various shapes are rendered into a bitmap grid.  They're treated
as black & white.

Determining whether or not a file holds a shape table is a tricky
business, because it's possible to have a broken header so long as
the accompanying program doesn't actually try to use the broken
entries.  Well-formed tables are given a higher likelihood rating
than those that seem a little broken.  We can't do pefect detection,
but it'll be listed as an option in the file viewer so long as the
basic file structure is compatible.