Add support for "overwrite all", part 1
When a file is about to to overwritten by add/extract/copy, the GUI
asks the user if they want to continue.  This was handled with a
simple MessageBox that had yes/no/cancel options.  This is fine for
a file or two, but inconvenient when working with a large number of
files.

This adds storage of persistent choices to the progress update callback
function, so that it can handle "overwrite all" and "overwrite none",
and adds a new file overwrite query dialog.  The dialog is fully
functional but incomplete, as the information we'd like to show
(pathnames of both files, sizes, modification dates) isn't yet
provided by the lower-level framework code.

(issue #19)