Label Finder icons with decimal
Finder icons are shown in a column, with the index number on the
left.  The icon metadata appears in the Notes tab, with the same
index value.  Showing this as a decimal value feels beter than
showing it as hex.

Fonts and shape-table grid labels remain in hex.  There's a good
argument to be made that shape tables should also be shown in decimal,
but hex labels work better for a 2D grid.