Add a warning for sparse block 0
ProDOS requires that the first block of the file be allocated, even
if it holds nothing but zeroes.  Most things will handle such files
just fine, but GS/OS and some P8 utilities can have difficulties.

The volume scan now adds warnings to the disk volume.  The files
themselves aren't flagged as damaged or dubious because it's okay
for us to interact with them.

Added a test disk image with sapling and tree samples.