Update sparse-first-block test case for forks
Both forks of extended files should follow the rule noted in the
previous change: the first block of the file should always be
allocated, even if it's all zero.  The previous change correctly
finds the problem, but the warning message didn't specify which fork
of the file was affected.

This updates the warning and adds extended files to the test image.