Change how file archives report read-only status
The file archive update mechanism generates a new file, deletes the
old file, and renames the new one to match the old.  Strictly
speaking, the IArchive object isn't read-only if the underlying file
is, because it can generate a new file without a problem (though the
higher-level code will fail when it tries to delete the old archive).

Having IArchive report IsReadOnly==false is a bit myopic.  If the
IArchive object's data stream is read-only, we want to treat the
archive is read-only so we don't have to add explicit archive-only
checks elsewhere.  The intent of marking a file read-only is to
ensure that nothing tries to modify it, and we should respect that
at all levels.

While it's conceivable that somebody might want to generate a new
archive from a read-only archive by using the API to modify it
directly, rather than copying records, it's far from the common case,
and anyone wishing to do so can sub-class Stream with a pass-through
class that always reports CanWrite=true (but rejects Write calls).