Add "simple block image" option
When creating a new disk image, or saving the contents of a disk or
partition to a new disk image, we offer all of the file formats that
are valid for the disk geometry.  One of the most common, the
unadorned block image, is presented as "Unadorned ProDOS-order (.po)",
which is fine for experienced Apple II users but perhaps a bit of a
mystery for everyone else.

We now offer "Simple block image (.iso/.hdv)", which does exactly
the same thing as the ProDOS block option but accepts different
filename extensions.