Add keyboard shortcuts for "recents"
The File > Recents menu is generated from the list of recently-opened
archives when the menu is opened.  To allow keyboard shortcuts to
work through the standard Command mechanism, we need to change that
to be less dynamic.  The menu is still generated as needed, but now
have a fixed set of six commands with keyboard shortcuts.

The list of recents is updated whenever an archive is opened, which
has some interesting effects.  Ctrl+Shift+1 reopens the current
archive, which isn't hugely useful, but also it works when nothing is
open.  Ctrl+Shift+2 opens the previous archive, allowing you to flip
between two archives quickly (which would admittedly be more useful
if you could copy & paste without needing to keep the source archive
open).  Ctrl+Shift+3 rotates between three archives.

Also, added keyboard shortcuts for Create Directory (Ctrl+Shift+N,
same as Windows explorer) and Add Files (Ctrl+Shift+A, because Ctrl+A
is already used for "select all").

(issue #25)