Fix some Stream subclasses
When implementing a subclass of Stream, which is IDisposable, the
overridden Dispose(bool) method must call base.Dispose(bool).

The NuLZWStream Read() method was assuming that the underlying
stream would return the full amount requested pre-EOF.  This is
currently valid but not guaranteed by the API.

This change should not cause any changes in behavior.