Add get-attr and rework set-attr
Multiple changes:
 - The order of arguments to set-attr has been changed, so that the
   attributes come before the filename.
 - set-attr now takes multiple filenames, and accepts wildcards and the
   --recurse argument.
 - The set-attr attribute changes are now provided in a comma-separated
   list, rather than space-separated.
 - Added new "get-attr" command.  This replaces the previous "show
   attributes if no edit args were specified" approach.

This was a bit awkward for MacZip archives.  set-attr is a read-
modify-write operation, but IArchive doesn't allow you to extract
file parts while a transaction is in progress.  So we need to
pre-flight all of the edits.

Documentation and tests have been updated.

Also, added a new line to the get-attr output that indicates whether
the entry is a plain file or a directory.  It also shows "dubious"
and "damaged" state.

(issue #30)