Generate HTML for disk/file format docs
The major search engines don't scan the contents of github repositories,
so to make these documents visible to web search we need to convert
the Markdown files to HTML and link to those.

This adds a single, flattened directory of documents, and updates the
documentation index page to point to them instead of inside the tree.

The conversions were created through the github REST API, via the
phseiff/github-flavored-markdown-to-html project.

This also updates the file encoding on MagicWindow-notes.md, which
has some non-ASCII characters and wasn't stored as UTF-8.