Audio recording support, part 1
This adds support for audio recordings made via the Apple II cassette
I/O port.  For convenience, we treat it as an archive format.  This
change has some basic WAV file parsing and a skeletal archive
implementation, as well as a notes file with information on the
cassette audio format.