Improve keyboard navigation
The Return key was mapped to View Files, but it's better if it acts
the same as double-click, which descends into directories and opens
embedded archives and disk images.

We now put the focus back on the file list after moving to a new
directory or moving up a level.  This allows keyboard navigation
through the directory tree while in single-dir view mode.

(issue #25)