Handle $C1/0000 > 32KB
A project under development is storing additional data in the
uncompressed PIC format, which is normally 32768 bytes.  IIgs
applications such as SuperConvert simply ignore everything past the
first 32KB.  We now do the same thing.

(I would argue that APF is a better way to handle this, but there's
no harm in leniency here, especially when it matches the behavior
of vintage applications.)