Make filename extensions on exports optional
Exported files are converted to a new format, such as PNG or CSV,
so they usually need a new filename extension.  This is not the case
when exporting source code files, which already have a meaningful
filename extension (.c, .asm, etc.), so adding ".txt" is unnecessary
and counter-productive.

This adds the "--[no-]add-ext" flag to the CLI, and the "Add Filename
Ext to Exports" checkbox to the GUI.  Both default to true.

This affects the export operation and the export-mode drag & drop
operation.  "Host" exports (e.g. exporting a GIF) are unaffected,
since those did not have an extension added previously.