Don't show Mac ZIP directory pairs in maczip mode
For some reason the "is it a Mac ZIP pair" function wasn't considering
ZIP directory entries as pairable, so we weren't excluding dir pairs.
They were also explicitly not tested by the CLI catalog/list routine.
This caused a bunch of "__MACOSX/.../" files to appear in the list.

This is only a problem when the __MACOSX entry is a directory, i.e.
a zero-length file ending with '/'.  The archives created by the
Mac Finder have an actual AppleDouble entry for directories.  It's
unclear what utility creates the zero-length pairing (possibly the
files were extracted to a non-Mac system and then re-zipped, and the
second zip action generated the directory entries).