Add DART disk image support, part 3
Implemented "best" compression (LZHUF).  This uses the common LZHUF
implementation, but without the leading length word, and with the
buffer initialization byte value changed from 0x20 to 0x00.

Updated the file analyzer to allow a third possibility, so that we
can add DART as an expected .image/.img file.

Minor issue: when DART files are stored in disk images, we can rely
on the HFS file type, but in ZIP archives we're currently not using
the type info from the MacZip record.  The result is that images
stored in ZIP archives without a meaningful extension aren't always
recognized as DART when the entry is double-clicked in the application.

Added an image without an extension to the test suite ZIP file, so
we have something to test the app's double-click behavior against.

Added another test file, this one with a bunch of uncompressible data,
to test the behavior when compression fails.  The tag data for all
blocks has been filled with random junk.