Add a debug info dump command
This adds an optional DebugDump() interface to filesystem and file
archive objects.  It can be used to output detailed human-readable
information about a file for debugging purposes.  Currently this is
only implemented for HFS (which had the debug code but nothing to
call it).

The interface is invoked by the new "debug-show-info" command.