Support text color in AWGS WP documents
The foreground color is specified as a palette entry, but the
document is drawn in 640 mode, which means four adjacent pixels use
different parts of the color palette.  We now generate a dithered
average of the pixel values and use that.  The results are a
reasonable approximation of what AWGS shows on screen.