Alter max allowed ANI file size
Since the converter is just converting the first frame, there's no
reason to care about the maximum file size.  That only matters if
we're trying to analyze the stream or to convert to GIF.  We now
do the ANI -> bitmap conversion for files of any size.  The extra
info added to the Notes tab is only provided for files <= 16MB.

The ANI -> animated GIF conversion didn't check the max file size
at all.  We now limit it to 512MB, which may not work in a 32-bit
environment, but this is an "experimental" feature so it's okay if
things are less than perfect.