Kluge ZX0 compression into NuFX
This uses the --set-int option to override the compression method,
which is usually LZW/2.  This value allows the user to specify any
supported compression format by number.  For example, to compress
files with ZX0, use "--set-int=nufx-comp-alg:8".

I enabled Deflate (6), because doing so was a three-line change that
can't possibly end in tears, but we don't support bzip2 (7).

This also adds ZX0 to the bulk compression test in the GUI.  It's
a couple orders of magnitude slower than anything else, though, so
I haven't tested a huge number of files.