Play with file viewer bitmap scaling
The behavior for unscaled converted bitmaps is unchanged.  We're
sticking with NearestNeighbor for now because the fuzziness of the
other options is annoying.

Upscaled bitmaps are now scaled manually.  This was mostly done as
an experiment with HighQuality scaling to see if we could reduce the
fuzziness at zoom > 1X (it does), but other issues arose.

(See issue #8)