Generate dirs with nonzero access
When copying & pasting entries from a file archive, the side that
initiates the copy generates fake directory entries.  This is useful
in situations where the clipboard recipient doesn't automatically
create subdirs as needed.

The directories were being created with access $00 (no permissions),
which caused problems down the road when the paste target was a
ProDOS disk image.

(issue #56)