Add "get-attr filename"
Useful for getting the volume name from a disk image.  For a DOS 3.x
disk, this returns the volume number as a plain decimal value.

(This might be better handled with a "volume info" command that returned
various disk image / file archive properties in machine-readable format,
like the filesystem type and sector ordering.)